+++ src/api/photoshop.ts
@@ -3,3 +3,13 @@ import { photoshop } from "../globals";
 export const notify = async (message: string) => {
   await photoshop.app.showAlert(message);
 };
+
+export const getProjectInfo = async () => {
+  const doc = photoshop.app.activeDocument;
+  const info = {
+    name: doc.name,
+    path: doc.path,
+    id: doc.id,
+  };
+  return info;
+};