+++ webview-ui/src/index-webview-vue.ts
@@ -1,0 +1,6 @@
+import { createApp } from "vue";
+import App from "./main-webview.vue";
+import "./app.css";
+import "./index.scss";
+
+createApp(App).mount("#app");